Moved new tests into corresponding files
